wip: #60 moved prisma client file to new lib folder
