+++ apps/issues/server/router/context.ts
@@ -5,7 +5,7 @@ import {
 import * as trpc from '@trpc/server';
 import * as trpcNext from '@trpc/server/adapters/next';
 
-import { prisma } from '../db/client';
+import { prisma } from '../../lib/db';
 import { authOptions as nextAuthOptions } from '../../pages/api/auth/[...nextauth]';
 
 type CreateContextOptions = {
